% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_utilities.R
\name{acquire_openeo}
\alias{acquire_openeo}
\title{Utility Function to Acquire Sentinel-2 Imagery using openEO}
\usage{
acquire_openeo(
  aoi,
  from_date,
  to_date,
  output_dir = tempdir(),
  scale_factor = 10000
)
}
\arguments{
\item{aoi, }{{sf} object, POLYGON or MULTIPOLYGON of area of interest}

\item{from_date, }{string, represents start of date range,
formatted as "YYYY-MM-DD"}

\item{to_date, }{string, end of date range, formatted as "YYYY-MM-DD"}

\item{output_dir, }{string, path to save downloaded, and processed imagery}

\item{scale_factor, }{integer, scaling factor for EO data source default 10000
, to scale Sentinel-2 15 bit DN to range (0, 1)}
}
\value{
list of BOA files
}
\description{
This non-exported function uses the \code{openeo} package
to send a request to Copernicus DataSpace, and prepare the products.
Called by \code{\link[rOPTRAM]{optram_acquire_s2}}
}
\note{
This function utilizes the \code{openeo} package.
Instructions for the login process:
First of all, to authenticate your account on the backend of the Copernicus
Data Space Ecosystem, it is necessary for you to complete the registration
process. Follow these instructions for registration:
\url{https://documentation.dataspace.copernicus.eu/Registration.html}
After you have registered and installed the \code{openeo} package, you can run the
\code{acquire_openeo} function.
During the process of connecting to the server and logging in, you need to
follow these steps:
\enumerate{
\item When the message "Press \if{html}{\out{<enter>}} to proceed:" appears in the console, press enter.
\item When prompted with 'Copy CTGB-UGFU and paste when requested by the browser' in the console, it may appear but can be ignored, as it is related to an older version of the openeo package.
\item Calling this method opens your system web browser, with which you can authenticate yourself on the back-end authentication system.
\item After that, the website will give you instructions to go back to the R client, where your connection has logged your account in. This means that every call that comes after that via the connection variable is executed by your user account.
\item You will be redirected to \url{https://identity.dataspace.copernicus.eu/}. Ensure you have an account and are logged in. You will be required to grant access - press "yes".
}
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2019-04-30"
aoi <- sf::st_read(system.file("extdata",
                "lachish.gpkg", package = 'rOPTRAM'))
Important! Assign the output to an object for later retrieval
acq <- acquire_openeo(aoi, from_date, to_date,
                      veg_index = "SAVI",
                      SWIR_band = 11)
}
}
