% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_s2.R
\name{acquire_s2_images}
\alias{acquire_s2_images}
\title{Acquire Sentinel 2 images at a given location and date range}
\usage{
acquire_s2_images(aoi, from_date, to_date, output_dir = tempdir())
}
\arguments{
\item{aoi:}{sf object, polygon boundary of area of interest}

\item{from_date:}{string, represents start of date range, formatted as "YYYY-MM-DD"}

\item{to_date:}{string, end of date range, formatted as "YYYY-MM-DD"}
}
\value{
output_path: string, path to downloaded files
}
\description{
Acquire Sentinel 2 images at a given location and date range
}
\examples{
print("acquire_s2_images.R")

}
