% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{aoi_to_name}
\alias{aoi_to_name}
\title{Get name string for Area of Interest from full file name}
\usage{
aoi_to_name(aoi_file)
}
\arguments{
\item{aoi_file, }{string, full path to AOI file}
}
\value{
aoi_name, string
}
\description{
Extract a string from the full path to AIO file
}
\examples{
aoi_file <- system.file("extdata", "migda_aoi.gpkg")
aoi_name <- aoi_to_name(aoi_file)
aoi_name
}
