% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilmoisture.R
\name{calculate_soil_moisture}
\alias{calculate_soil_moisture}
\title{Calculate soil moisture grid}
\usage{
calculate_soil_moisture(img_date)
}
\arguments{
\item{img_date:}{string, image date of Sentinel 2 acquisition}
}
\value{
rast, soil moisture grid
}
\description{
Calculate soil moisture grid
}
\examples{
print("Running calculate_soil_moisture.R")
}
