% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_str}
\alias{calculate_str}
\title{Calculate STR from SWIR band of bottom of atmosphere images}
\usage{
calculate_str(img_stk, swirband = 5)
}
\arguments{
\item{img_stk, }{terra SpatRaster, multiband stack of images, already clipped to aoi}

\item{swirband, }{integer, number of red band}
}
\value{
STR, SpatRaster of STR band
(not exported)
}
\description{
Use this function to prepare STR from SAFE imagery
when you have already downloaded Sentinel 2 image files in advance
(without using \code{sen2r}).
}
