% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_vi}
\alias{calculate_vi}
\title{Calculate NDVI, SAVI or MSAVI from bottom of atmosphere images}
\usage{
calculate_vi(
  img_stk,
  viname = "NDVI",
  redband = 3,
  greenband = 2,
  blueband = 1,
  nirband = 4
)
}
\arguments{
\item{img_stk, }{terra SpatRaster, multiband stack of images,
already clipped to aoi}

\item{viname, }{string, which VI to prepare,
either 'NDVI' (default) or 'SAVI' or 'MSAVI'}

\item{redband, }{integer, number of red band}

\item{greenband, }{integer, number of green band}

\item{blueband, }{integer, number of blue band}

\item{nirband, }{integer, number of NIR band}
}
\value{
vi_rast, SpatRaster of vegetation index
}
\description{
Use this function to prepare vegetation index from SAFE imagery
when you have already downloaded Sentinel 2 image files in advance
(without using \code{sen2r}).
}
\examples{
img_stk <- terra::rast(system.file("extdata",
         "BOA",
         "S2A2A_20230301_121_migdaaoi_BOA_10.tif",
         package = "rOPTRAM"))
vi <- calculate_vi(img_stk)
}
