% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_str.R
\name{CalculateSTR}
\alias{CalculateSTR}
\title{Create SWIR transformed reflectance}
\usage{
CalculateSTR(BOA_dir)
}
\arguments{
\item{BOA_dir}{}
}
\value{
list of string, the path to transformed raster
}
\description{
Create SWIR transformed reflectance
}
\examples{
print("Running calculate_str.R")
}
