% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_utilities.R
\name{check_gcloud}
\alias{check_gcloud}
\title{Check if gcloud is installed}
\usage{
check_gcloud()
}
\value{
boolean
}
\description{
Check for sen2r dependency, and gcloud utility
}
\examples{
\dontrun{
gcloud_ok <- check_gcloud()
}

}
