% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_scihub_access}
\alias{check_scihub_access}
\title{Check access to Sentinel Hub, using sen2r}
\usage{
check_scihub_access()
}
\value{
boolean, whether access to scihub using sen2r is possible
}
\description{
Verify that sen2r is installed,
}
\examples{
check_scihub_access()
}
