% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_landsat.R
\name{crop_landsat_list}
\alias{crop_landsat_list}
\title{Crop List of Landsat Bands to AOI}
\usage{
crop_landsat_list(landsat_list)
}
\arguments{
\item{landsat_list, }{list of char, full paths landsat images}
}
\value{
cropped_list, list, paths to derived BOA stacks, cropped to study area
}
\description{
Utility function to prepare BOA bands cropped to Area of Interest
}
\examples{
\dontrun{
 landsat_list <- list.dirs(landsat_dir)[
                grepl("L*_02_*",list.dirs(landsat_dir))]
cropped_landsat_list <- crop_landsat_list(landsat_list)
}
}
