% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{exponential_coefficients}
\alias{exponential_coefficients}
\title{Utility Function to Prepare Exponential FittedEedges of Trapezoid}
\usage{
exponential_coefficients(df, output_dir)
}
\arguments{
\item{df, }{data.frame, values of VI and STR along edges of trapezoid}

\item{output_dir, }{string, path to save coefficients CSV file}
}
\value{
rmse_list, list of RMSE values of trapezoid fitted lines
for both wet and dry edges
}
\description{
Called by \code{optram_wetdry_coefficients()}
to prepare exponential curve along trapezoid edges
Calculates the intercept and slope of both wet and dry edges
and updates the edges data.frame with these exp fitted values
Not exported
}
\examples{
\dontrun{
  df <- read.csv(system.file("extdata", "trapezoid_edges.csv",
                package = "rOPTRAM"))
  output_dir <- tempdir()
  coeffs <- rOPTRAM::exponential_coefficients(df, output_dir)
  coeffs
}
}
