% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{exponential_soil_moisture}
\alias{exponential_soil_moisture}
\title{Calculate Soil Moisture Grid, Exponential Trapezoid}
\usage{
exponential_soil_moisture(coeffs, VI, STR)
}
\arguments{
\item{coeffs, }{list, 4 trapezoid coefficients}

\item{VI, }{terra rast, the vegetation index raster}

\item{STR, }{terra rast, the STR raster}
}
\value{
rast, soil moisture grid
}
\description{
Prepare soil moisture grid from STR and VI images
for a single date, based on exponential function
to derive trapezoid edges,
using the derived slope and intercept coefficients
}
\note{
This function is used after preparing the OPTRAM model coefficients with:
\code{optram_wetdry_coefficients()}. Typically a new image date,
(that was not used for preparing the model),
will be referenced in the \code{img_date} parameter.
The resulting soil moisture raster is saved to \code{output_dir}.
This function implements an exponential trapezoid, following:
Ambrosone, Mariapaola, et al. 2020.
“Retrieving Soil Moisture in Rainfed and Irrigated Fields
Using Sentinel-2 Observations and a Modified OPTRAM Approach.”
International Journal of Applied Earth Observation and Geoinformation 89 (July):
102113. https://doi.org/10.1016/j.jag.2020.102113.
}
\examples{
img_date <- "2023-03-11"
VI_dir <- system.file("extdata", "NDVI", package = "rOPTRAM")
STR_dir <- system.file("extdata", "STR", package = "rOPTRAM")
SM <- optram_calculate_soil_moisture(img_date,
            VI_dir, STR_dir,
            data_dir = system.file("extdata"))
}
