% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{linear_soil_moisture}
\alias{linear_soil_moisture}
\title{Calculate Soil Moisture Grid, Linear Trapezoid}
\usage{
linear_soil_moisture(coeffs, VI, STR)
}
\arguments{
\item{coeffs, }{list, 4 trapezoid coefficients}

\item{VI, }{SpatRaster, the vegetation index raster}

\item{STR, }{SpatRaster, the STR raster}
}
\value{
rast, soil moisture grid
}
\description{
Prepare soil moisture grid from STR and VI images
for a single date, using linear trapezoid method, and
using the derived slope and intercept coefficients
}
\note{
This function is used after preparing the OPTRAM model coefficients with:
\code{optram_wetdry_coefficients()}. Typically a new image date,
(that was not used for preparing the model),
will be referenced in the \code{img_date} parameter.
The resulting soil moisture raster is saved to \code{output_dir}.
}
\examples{
img_date <- "2023-03-11"
VI_dir <- system.file("extdata", "NDVI", package = "rOPTRAM")
STR_dir <- system.file("extdata", "STR", package = "rOPTRAM")
SM <- optram_calculate_soil_moisture(img_date,
            VI_dir, STR_dir,
            data_dir = system.file("extdata"))
}
