% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram.R
\name{optram}
\alias{optram}
\title{Prepare Sentinel imagery for soil moisture prediction using OPTRAM.}
\usage{
optram(aoi, vi = "NDVI", output_dir = tempdir())
}
\arguments{
\item{aoi:}{string, full path to polygon spatial file of area of interest}

\item{vi:}{string, which VI to prepare, either 'NVDI' or 'SAVI'}

\item{coeffs_file:}{string, where to save coeffs_file, default is tempdir()}
}
\value{
coeffs_files: string, full path to saved CSV of wet-dry coefficients
}
\description{
The main wrapper function to download, and preprocess Sentinel 2 imagery
then prepare the OPTRAM trapezoid to derive slope and intercept
for wet and dry trapezoid lines.
See:
Sadeghi, M., Babaeian, E., Tuller, M., Jones, S.B., 2017.
The optical trapezoid model:
A novel approach to remote sensing of soil moisture
applied to Sentinel-2 and Landsat-8 observations.
Remote Sensing of Environment 198, 52–68,
https://doi.org/10.1016/j.rse.2017.05.041 .
}
\examples{
print("Running optram_prepare_coeffs.R")
}
