% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_soilmoisture.R
\name{optram_calculate_soil_moisture}
\alias{optram_calculate_soil_moisture}
\title{Calculate Soil Moisture Grid}
\usage{
optram_calculate_soil_moisture(
  img_date,
  VI_dir,
  STR_dir,
  coeffs_file,
  output_dir = tempdir(),
  trapezoid_method = c("linear", "exponential", "polynomial")
)
}
\arguments{
\item{img_date, }{string, image date of single Sentinel 2 acquisition
formatted as "YYYY-MM-DD"}

\item{VI_dir, }{string, full path to directory holding the VI rasters}

\item{STR_dir, }{string, full path to directory holding the STR rasters}

\item{coeffs_file, }{string, full path to csv file of wet-dry coefficients}

\item{output_dir, }{string, full path to output directory
for saving soil moisture raster}

\item{trapezoid_method, }{string,
either "linear", "exponential", or "polynomial", default is "linear"}
}
\value{
rast, soil moisture grid
}
\description{
Prepare soil moisture grid from STR and VI images
for a single date
using the derived slope and intercept coefficients
}
\note{
This function is used after preparing the OPTRAM model coefficients with:
\code{optram_wetdry_coefficients()}. Typically a new image date,
(possibly an image that was not used for preparing the model),
will be referenced in the \code{img_date} parameter.
The resulting soil moisture raster is saved to \code{output_dir}.

Three trapezoid models are offered through the trapezoid_method parameter:
either "linear", "exponential", or "polynomial".
For further details see:
Ambrosone, Mariapaola, et al. 2020.
“Retrieving Soil Moisture in Rainfed and Irrigated Fields
Using Sentinel-2 Observations and a Modified OPTRAM Approach.”
International Journal of Applied Earth Observation and Geoinformation 89 (July):
102113. https://doi.org/10.1016/j.jag.2020.102113.

The coeffs_file string must be set to match the trapezoid_method.
i.e. for "linear" method the file was saved as 'coefficients.csv'
for "exponential" it was saved as 'coefficients_exp.csv'
for "polynomial" it was saved as 'coefficients_poly.csv'
}
\examples{
img_date <- "2023-03-11"
VI_dir <- system.file("extdata", "SAVI", package = "rOPTRAM")
STR_dir <- system.file("extdata", "STR", package = "rOPTRAM")
coeffs_file <- system.file("extdata", "coefficients.csv",
        package = "rOPTRAM")
SM <- optram_calculate_soil_moisture(img_date, VI_dir, STR_dir,
                                     coeffs_file,
                                     trapezoid_method = "linear")
}
