% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_ndvi_str.R
\name{optram_ndvi_str}
\alias{optram_ndvi_str}
\title{Prepare dataframe of pairs NDVI and STR values}
\usage{
optram_ndvi_str(STR_list, VI_list, output_dir = tempdir())
}
\arguments{
\item{STR_list, }{list of paths to STR raster files}

\item{VI_list, }{list of paths to NDVI raster files}

\item{output_dir, }{string, path to save data.frames (in RDS format)}
}
\value{
full_df, data.frame with 5 columns: X,Y,Date,NDVI,STR
}
\description{
Collect all pixel values of both vegetation index,
and Swir Transformed Reflectance, for a time series of images.
Prepare data.frame of all pairs of values
(as input for the \code{optram_wetdry_coefficients()} function)
}
\examples{
print("Running optram_ndvi_str.R")
}
