% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{optram_options}
\alias{optram_options}
\title{Display or Set Package Options}
\usage{
optram_options(opt_name = NULL, opt_value = NULL, show_opts = TRUE)
}
\arguments{
\item{opt_name, }{string, one of the package options.
Default NULL}

\item{opt_value, }{string, numeric, or boolean, depending on opt_name.}

\item{show_opts, }{boolean, default TRUE, prints a list of all optram options
Default NULL}
}
\description{
{rOPTRAM} uses several package options.
This function displays the current defined options,
(showing the default options when the package is first loaded),
and allows users to set new values for each option.
}
\note{
When no new option name or value is specified,
a list of currently defined options is printed.

\code{rOPTRAM} defines the following options at startup
\tabular{lll}{
opt_name     \tab default  \tab other possible values \cr
---------------- \tab---------------\tab ------------------- \cr
veg_index  \tab "NDVI"  \tab "SAVI", "MSAVI" \cr
remote     \tab "scihub"   \tab "openeo" \cr
period     \tab "full"   \tab "seasonal" \cr
max_cloud  \tab 12 \tab between 0 and 100 \cr
vi_step    \tab 0.005  \tab usually between 0.01 and 0.001 \cr
trapezoid_method \tab "linear"  \tab "polynomial" or "exponential" \cr
SWIR_band  \tab 11    \tab 12 \cr
max_tbl_size \tab 1e+6  \tab depends on computer resources, default 1e6, minimum: 1e4 \cr
rm.low.vi   \tab FALSE \tab TRUE \cr
rm.hi.str   \tab FALSE \tab TRUE \cr
plot_colors \tab "no"  \tab "no" = uniform green color for all points \cr
\tab  \tab "features" = points colored by aoi features \cr
\tab  \tab "density" = points colored by point density \cr
\tab  \tab "contours" = plots density contour lines \cr
\tab  \tab "months" = points colored by month of image date \cr
feature_col \tab "ID" \tab string, name of \emph{numeric} column \cr
\tab  \tab that contains feature ID's for coloring plot \cr
edge_points \tab TRUE  \tab FALSE, whether to add \cr
\tab  \tab the trapezoid edge points to the plot \cr
only_vi_str \tab  FALSE \tab TRUE (avoids downloading all Sentinel bands) \cr
tileid      \tab  NA \tab a string of 5 characters, \cr
\tab \tab  download only the requested tileid  \cr
scm_mask    \tab TRUE \tab FALSE, whether to mask out: \cr
\tab \tab cloud shadows, clouds, water, snow, using the Copernicus SCM product. \cr
overwrite   \tab  FALSE \tab Set to TRUE to re-download previously acquired images. \cr
save_img_list   \tab  FALSE \tab Set to TRUE to save list of available images to RDS file. \cr
resolution   \tab  10 \tab Choose resolution of downloaded Sentinel images, 10m, 20m or 60m \cr
area_cover   \tab  99.0 \tab Use only images that cover at least area_cover \% of AOI \cr
porosity \tab 0.4 \tab Soil porosity value applied to OPTRAM output to get VWC \cr
}
}
\examples{
opts <- options()
optram_options()      # prints out list of current options
optram_options("SWIR_band", 12)
optram_options("veg_index", "SAVI")
optram_options("trapezoid_list", "exp")  # fails
options(opts)

}
