% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_other.R
\name{optram_prepare_other_vi_str}
\alias{optram_prepare_other_vi_str}
\title{Handle Third Party Imagery With Red, NIR and SWIR Bands}
\usage{
optram_prepare_other_vi_str(
  img_dir,
  aoi,
  viname = "NDVI",
  output_dir = tempdir()
)
}
\arguments{
\item{img_dir, }{string, full path directory of downloaded images
(assumes Geotiff)}

\item{aoi, }{string, full path to polygon spatial file of area of interest}

\item{viname, }{string, which VI to prepare, 'NDVI', 'SAVI', etc.}

\item{output_dir, }{string, where to save Geotiff, default is tempdir()}
}
\value{
output_files, list, full paths to saved Geotiff files
(not exported yet)
}
\description{
This function prepares vegetation index and STR rasters
from other (non-Sentinel) image files downloaded in advance.
}
