% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_safe.R
\name{optram_prepare_safe_vi_str}
\alias{optram_prepare_safe_vi_str}
\title{Handle Sentinel imagery in original Copernicus SAFE format}
\usage{
optram_prepare_safe_vi_str(safe_dir, aoi, vi = "NDVI", output_dir = tempdir())
}
\arguments{
\item{safe_dir:}{string, full path directory of downloaded Sentinel 2 images in SAFE format}

\item{aoi:}{sf object, polygon of area of interest}

\item{vi:}{string, which VI to prepare, either 'NVDI' or 'SAVI'}

\item{output_dir:}{string, where to save Geotiff, default is tempdir()}
}
\value{
output_files:list, full paths to saved Geotif files
}
\description{
This function prepares vegetation index and STR rasters
from Sentinel 2 image files downloaded in advance (without using \code{sen2r}).
}
\examples{
print("Running optram_prepare_safe_vi_str.R")

}
