% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{optram_wetdry_coefficients}
\alias{optram_wetdry_coefficients}
\title{Derive coefficients of slope and intercept}
\usage{
optram_wetdry_coefficients(
  full_df,
  output_dir = tempdir(),
  step = 0.001,
  save_plot = TRUE
)
}
\arguments{
\item{step, }{float}

\item{full_df:}{data.frame of STR and NDVI values}

\item{output_dir:}{string, directory to save coefficients CSV file}

\item{save_plot:}{boolean, If TRUE (default) save scatterplot to output_dir}
}
\value{
list of float, coefficients of wet-dry trapezoid
}
\description{
Derive slope and intercept coefficients
for both wet and dry trapezoid lines.
Write coefficients to a CSV file (as input to \code{optram_soilmoisture()} function)
}
\examples{
print("Running optram_wetdry_coefficients.R")

}
