% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{plot_cloud_polynomial}
\alias{plot_cloud_polynomial}
\title{Scatterplot, Polynomial Fitted Curve}
\usage{
plot_cloud_polynomial(pl_base, output_dir, aoi_name)
}
\arguments{
\item{pl_base, }{ggplot object, additional lines are added to base plot}

\item{output_dir, }{string, path to directory where trapezoid edges were saved}

\item{aoi_name, }{string, added to plot title}
}
\value{
ggplot object
}
\description{
Plot of VI/STR points, with polynomial fitted curve
}
\note{
This function is called from \code{plot_vi_str_cloud()}
}
\examples{
\dontrun{
output_dir <- system.file("extdata")
pl <- plot_cloud_polynomial(pl_base, output_dir)
}
}
