% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_validate.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot validation regression line}
\usage{
plot_correlation(sm_df, output_dir)
}
\arguments{
\item{sm_df, }{data.frame, values of in situ and model predicted soil moisture
over a range of dates}

\item{output_dir, }{string, full path to save png file of plot}
}
\value{
png_file, string, full path to png file of plot
}
\description{
Prepare a scatterplot of in situ measurements of soil moisture
against model predicted values at TDR TDRs.
and add regression line
}
\examples{
print("Running correlation_plot()")

}
