% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{plot_ndvi_str_cloud}
\alias{plot_ndvi_str_cloud}
\title{Create scatter plot of STR-NDVI point cloud,}
\usage{
plot_ndvi_str_cloud(full_df, coeffs, aoi_name, output_dir = tempdir())
}
\arguments{
\item{full_df, }{data.frame of NDVI and STR pixel values}

\item{coeffs, }{list of floats, the slope and intercept
of wet and dry regression lines}

\item{aoi_name, }{string, used in plot title}

\item{output_dir, }{string, directory to save plot png file.}
}
\value{
None
}
\description{
Plot STR-NDVI scatterplot to show dry and wet trapezoid lines
over scatterplot of multi-temporal STR and NDVI pixel values
}
\examples{
aoi_name <- "Test"
full_df <- readRDS(system.file("extdata", "VI_STR_data.rds", package = "rOPTRAM"))
coeffs <- read.csv(system.file("extdata", "coefficients.csv", package = "rOPTRAM"))
plot_ndvi_str_cloud(full_df, coeffs, aoi_name)
}
