% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{plot_vi_str_cloud}
\alias{plot_vi_str_cloud}
\title{Create Scatter Plot of STR-NDVI Data Points,}
\usage{
plot_vi_str_cloud(full_df, aoi_name, edges_df, output_dir = tempdir())
}
\arguments{
\item{full_df, }{data.frame of NDVI and STR pixel values}

\item{aoi_name, }{string, used in plot title}

\item{edges_df, }{data.frame, points along the wet/dry edges for trapezoid}

\item{output_dir, }{string, directory to save plot png file.}
}
\value{
None
}
\description{
Plot STR-NDVI scatterplot to show dry and wet trapezoid lines
over scatterplot of multi-temporal STR and NDVI pixel values
}
\examples{
aoi_name <- "Test"
full_df <- readRDS(system.file("extdata", "VI_STR_data.rds",
        package = "rOPTRAM"))
edges_df <- read.csv(system.file("extdata", "trapezoid_edges_lin.csv",
                        package = "rOPTRAM"))
plot_vi_str_cloud(full_df, aoi_name, edges_df)
edges_df <- read.csv(system.file("extdata", "trapezoid_edges_poly.csv",
                        package = "rOPTRAM"))
optram_options("trapezoid_method", "polynomial")
plot_vi_str_cloud(full_df, aoi_name, edges_df)

}
