% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{polynomial_coefficients}
\alias{polynomial_coefficients}
\title{Utility Function to Prepare Polynomial Fitted Edges of Trapezoid}
\usage{
polynomial_coefficients(df, output_dir)
}
\arguments{
\item{df, }{data.frame, values of VI and STR along edges of trapezoid}

\item{output_dir, }{string, path to save coefficients CSV file}
}
\value{
df, data.frame, the trapezoid line edge points with fitted values added
for both wet and dry edges
}
\description{
Called by \code{\link[rOPTRAM]{optram_wetdry_coefficients}} to prepare second order polynomial curve along trapezoid edges
Calculates six coefficients:  intercept (alpha)  of both wet and dry edges
and first and second order coefficients (beta), as in
\eqn{STR = alpha + beta_1 * VI + beta_2 * VI^2}
and updates the edges data.frame with these polynomila values fitted values
}
\note{
Three CSV files are saved:
\itemize{
\item the regressions coefficients,
\item the trapezoid edge points
\item RMSE of the fitted curve
}
}
\examples{
\dontrun{
  df <- read.csv(system.file("extdata", "trapezoid_edges.csv",
                  package = "rOPTRAM"))
  output_dir <- tempdir()

 }
}
