% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{print_edges_rmse}
\alias{print_edges_rmse}
\title{Print RMSE of Trapezoid Fitted Curve}
\usage{
print_edges_rmse(edges_file)
}
\arguments{
\item{trapezoid_edges, }{string, full path to  trapezoid_egdes file}
}
\value{
character vector, two values of RMSE for wet and dry edges
}
\description{
Calculate and print out RMSE for fitted trapezoid edges
}
\note{
Not exported,
RMSE values are printed from each of the coefficients functions
}
\examples{
edges_file <- system.file("extdata", "trapezoid_edges.csv",
                           package = "rOPTRAM")
df <- utils::read.csv(edges_file)
output_dir <- tempdir()
coeffs <- linear_coefficients(df, output_dir)

}
