% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{print_rmse}
\alias{print_rmse}
\title{Print RMSE of Trapezoid Fitted Curve}
\usage{
print_rmse(edges_file)
}
\arguments{
\item{trapezoid_edges, }{string, full path to  trapezoid_egdes file}
}
\value{
character vector, two values of RMSE for wet and dry edges
}
\description{
Calculate and print out RMSE for fitted trapezoid edges
}
\note{
Not exported
}
