% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{retrieve_cdse_credentials}
\alias{retrieve_cdse_credentials}
\title{Retrieve CDSE Client Credentials from File}
\usage{
retrieve_cdse_credentials()
}
\value{
A data frame containing the retrieved CDSE clientid and secret,
or NULL if credentials are not available.
}
\description{
Retrieve CDSE clientid and secret from file
The file location is system specific. It would have been setup
in advance using the \code{\link[rOPTRAM]{store_cdse_credentials}} function
}
