% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOPTRAM-package.R
\docType{package}
\name{rOPTRAM-package}
\alias{rOPTRAM}
\alias{rOPTRAM-package}
\title{rOPTRAM: Derive soil moisture using the OPTRAM algorithm}
\description{
The OPtical TRapezoid Model (OPTRAM) derives soil moisture based on the linear relation between a vegetation index and Land Surface Temperature (LST). The Short Wave Infra-red (SWIR) band is used as a proxy for LST. See: Sadeghi, M., Babaeian, E., Tuller, M., Jones, S.B., 2017. The optical trapezoid model: A novel approach to remote sensing of soil moisture applied to Sentinel-2 and Landsat-8 observations. Remote Sensing of Environment 198, 52–68, https://doi.org/10.1016/j.rse.2017.05.041 .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/rsl-bidr/roptram.git}
  \item Report bugs at \url{https://gitlab.com/rsl-bidr/}
}

}
\author{
\strong{Maintainer}: Micha Silver \email{silverm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-1128-1325}{ORCID})

Other contributors:
\itemize{
  \item Arnon Karnieli \email{karnieli@bgu.ac.il} (\href{https://orcid.org/0000-0001-8065-9793}{ORCID}) [contributor, funder]
}

}
\keyword{internal}
