% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{store_cdse_credentials}
\alias{store_cdse_credentials}
\title{Store CDSE Client Credentials}
\usage{
store_cdse_credentials(clientid = NULL, secret = NULL)
}
\arguments{
\item{clientid, }{string, user's OAuth client id}

\item{secret, }{string, user's OAuth secret}
}
\description{
Store CDSE clientid and secret into a file
The file location is system specific.
Users who chose to save CDSE credentials can use this function
(and then the \code{retrieve_cdse_credentials()} afterwards)
The clientid and secret are obtained from:
}
