% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_utilities.R
\name{acquire_openeo}
\alias{acquire_openeo}
\title{Utility Function to Acquire Sentinel-2 Imagery using openEO}
\usage{
acquire_openeo(
  aoi_file,
  from_date,
  to_date,
  max_cloud = 10,
  output_dir = tempdir(),
  veg_index = "NDVI",
  scale_factor = 10000
)
}
\arguments{
\item{aoi_file, }{string, full path to polygon spatial file of
boundary of area of interest}

\item{from_date, }{string, represents start of date range,
formatted as "YYYY-MM-DD"}

\item{to_date, }{string, end of date range, formatted as "YYYY-MM-DD"}

\item{max_cloud, }{integer, maximum percent of cloud cover. Default 10.}

\item{output_dir, }{string, path to save downloaded, and processed imagery}

\item{veg_index, }{string, which index to prepare. Default "NDVI".
Can be "NDVI", "SAVI", "MSAVI", etc}

\item{scale_factor, }{integer, scaling factor for EO data source default 10000
, to scale Sentinel-2 15 bit DN to range (0, 1)}
}
\value{
list of BOA files
}
\description{
This non-exported function uses the {openeo} package
to send a request to Copernicus DataSpace, and prepare the products.
Called by optram_acquire_s2
}
\note{
This function utilizes the \code{openeo} package.
Instructions for the login process:
First of all, to authenticate your account on the backend of the Copernicus
Data Space Ecosystem, it is necessary for you to complete the registration
process. Follow these instructions for registration:
https://documentation.dataspace.copernicus.eu/Registration.html
After you have registered and installed the \code{openeo} package, you can run the
\code{acquire_openeo} function.
During the process of connecting to the server and logging in, you need to
follow these steps:
A. When the message "Press \if{html}{\out{<enter>}} to proceed:" appears in the console,
press enter.
Calling this method opens your system web browser, with which
you can authenticate yourself on the back-end authentication system. After
that, the website will give you instructions to go back to the R client,
where your connection has logged your account in. This means that every
call that comes after that via the connection variable is executed by your
user account.
B. You will be redirected to "https://identity.dataspace.copernicus.eu/".
Ensure you have an account and are logged in. You will be required to
grant access - press "yes".
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2019-04-30"
aoi <- system.file("extdata", "migda_aoi.gpkg", package = 'rOPTRAM')
acquire_openeo(aoi, from_date, to_date,
               veg_index = "SAVI")
}
}
