% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_utilities.R
\name{acquire_scihub}
\alias{acquire_scihub}
\title{Utility Function to Acquire Sentinel-2 Imagery Using
the {CDSE} Package}
\usage{
acquire_scihub(
  aoi_file,
  from_date,
  to_date,
  max_cloud = 10,
  output_dir = tempdir(),
  veg_index = "NDVI",
  save_creds = TRUE,
  clientid = NULL,
  secret = NULL,
  SWIR_band = c(11, 12)
)
}
\arguments{
\item{aoi_file, }{string, full path to polygon spatial file of
boundary of area of interest}

\item{from_date, }{string, represents start of date range,
formatted as "YYYY-MM-DD"}

\item{to_date, }{string, end of date range, formatted as "YYYY-MM-DD"}

\item{max_cloud, }{integer, maximum percent of cloud cover. Default 10.}

\item{output_dir, }{string, path to save downloaded, and processed imagery}

\item{veg_index, }{string, which index to prepare. Default "NDVI".
Can be "NDVI", "SAVI", "MSAVI", etc}

\item{save_creds, }{logical, whether to save CDSE credentials. Default TRUE.}

\item{clientid, }{string, user's OAuth client id. Required if \code{save_creds}
is TRUE.}

\item{secret, }{string, user's OAuth secret. Required if \code{save_creds} is TRUE.}

\item{SWIR_band, }{integer, either 11 or 12, determines which SWIR band to use}
}
\value{
list of BOA files
}
\description{
This non-exported function uses the {CDSE} package
to send a request to Copernicus Dataspace, and prepare the products.
Called by optram_acquire_s2()
}
\note{
This function utilizes the \code{CDSE} package.
Make sure to install the CDSE and jsonlite packages.
Create OAuth account and token:
Creating an Account:
\enumerate{
\item Navigate to the \href{https://dataspace.copernicus.eu/}{Copernicus portal}.
\item Click the "Register" button to access the account creation page.
\item If already registered, enter your username and password,
and click "Login."
\item Once logged in, go to the User dashboard and click "User Settings" to
access the Settings page.
}

Creating OAuth Client:
\enumerate{
\item On the Settings page, click the green "Create New" button located on
the right.
\item Enter a suitable "Client Name" and click the green "Create Client"
button.
\item A Client secret is generated.
the user must save her secret and clientid somewhere.
these credentials will be saved automatically to a standard filesystem
location if the user calls check_scihub() with the argument save_creds
set to TRUE (recommended).
if the user chooses not to save credentials to the standard filesystem
location, then she will need to add both clientid and secret to each
acquire_scihub() function call.
}

Using Credentials with \code{aquire_scihub}:
\itemize{
\item Now, you can utilize the generated \code{clientid} and \code{secret} in
the \code{aquire_scihub} function.
\item If you want to store your credentials on your computer, ensure that when
running \code{aquire_scihub}, the \code{save_creds} parameter is set to \code{TRUE}.
\item During the first run of \code{aquire_scihub}, manually input your \code{clientid}
and \code{secret} in the function's signature. Subsequent runs will use the
stored credentials.
}

Subject Area Constraint:
The downloadable images are restricted to a maximum size of 2500 pixels on
each side. This limitation is established due to the final resolution set to
10 meters using JavaScript. Consequently, the subject area available for
download is limited to 25 kilometers in both directions. Please be aware of
this restriction when selecting your desired area for download.
Area of Interest (AOI) Specification:
When defining your Area of Interest (AOI), please ensure that it is
represented as a polygonal layer with only one feature. This feature can
either be a single POLYGON or a MULTIPOLYGON, which may consist of
non-contiguous areas, but only one feature is permissible.
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2019-04-30"
aoi <- system.file("extdata", "migda_aoi.gpkg", package = 'rOPTRAM')
acquire_scihub(aoi, from_date, to_date,
               veg_index = "SAVI")
}
}
