% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_str}
\alias{calculate_str}
\title{Calculate STR from SWIR Bottom of Atmosphere Band}
\usage{
calculate_str(img_stk, SWIR_band = c(11, 12), scale_factor = 10000)
}
\arguments{
\item{img_stk, }{terra SpatRaster, multiband stack of images,
already clipped to aoi}

\item{SWIR_band, }{integer, number of SWIR band, one of (11 or 12)}

\item{scale_factor, }{integer, scaling factor for EO data source
default 10000, to scale Sentinel-2 15 bit DN to range (0, 1)}
}
\value{
STR, SpatRaster of STR band
}
\description{
Use this function to prepare STR from SAFE imagery
when you have already downloaded Sentinel 2 image files in advance
}
\note{
For Landsat images, scale_factor should be 1,
since Landsat metadata contains gain and offset
for scaling image bands.
}
\examples{
img_stk <- terra::rast(system.file("extdata", "BOA",
         "BOA_2022-12-11.tif", package = "rOPTRAM"))
str <- calculate_str(img_stk, STR_dir = tempdir(), SWIR_band = 11)
}
