% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_vi}
\alias{calculate_vi}
\title{Calculate NDVI, SAVI or MSAVI from bottom of atmosphere images}
\usage{
calculate_vi(img_stk, vi = "NDVI", redband = 3, nirband = 4)
}
\arguments{
\item{img_stk, }{terra SpatRaster, multiband stack of images, already clipped to aoi}

\item{vi, }{string, which VI to prepare, either 'NDVI' (default) or 'SAVI' or 'MSAVI'}

\item{redband, }{integer, number of red band}

\item{nirband, }{integer, number of NIR band}
}
\value{
vi_rast, SpatRaster of vegetation index
}
\description{
Use this function to prepare vegetation index from SAFE imagery
when you have already downloaded Sentinel 2 image files in advance
(without using \code{sen2r}).
}
