% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_safe.R
\name{calculate_vi}
\alias{calculate_vi}
\title{Calculate NDVI or SAVI from bottom of atmosphere images}
\usage{
calculate_vi(img_stk, vi = "NDVI", output_dir = tempdir())
}
\arguments{
\item{red:}{terra SpatRaster of red band, already clipped to AOI}

\item{nir:}{terra SpatRaster of nir band, already clipped to AOI}

\item{vi:}{string, which VI to prepare, either 'NVDI' (default) or 'SAVI' or 'MSAVI'}

\item{output_dir:}{string, where to save VI and STR and Geotiff, default is tempdir()}
}
\value{
output_files:list, full paths to saved Geotif files
(not exported)
}
\description{
Use this function to prepare vegetation index from SAFE imagery
when you have already downloaded Sentinel 2 image files in advance
(without using \code{sen2r}).
}
\examples{
print("Running optram_prepare_safe_vi_str.R")

}
