% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_vi}
\alias{calculate_vi}
\title{Calculate Vegetation Index from Bottom of Atmosphere Image Bands}
\usage{
calculate_vi(
  img_stk,
  redband = 4,
  greenband = 3,
  blueband = 2,
  nirband = 5,
  scale_factor = 2^15
)
}
\arguments{
\item{img_stk, }{terra SpatRaster, multiband stack of images,
already clipped to aoi}

\item{redband, }{integer, number of red band}

\item{greenband, }{integer, number of green band}

\item{blueband, }{integer, number of blue band}

\item{nirband, }{integer, number of NIR band}

\item{scale_factor, }{numeric, scales Sentinel-2 15 bit values
down to range (0,255) as expected for vegetation indices
default 32768 (2^15)}
}
\value{
vi_rast, SpatRaster of vegetation index
}
\description{
Use this function to prepare vegetation index from SAFE imagery
when you have already downloaded Sentinel 2 image files in advance
}
\note{
The scale_factor parameter reduces numeric range "digital number",
(DN) of Sentinel-2 images from 15 bit integer to 8 bit.
Vegetation indices such as SAVI expect values in the range (0, 255).
This scale_factor produces values in that range.
When using Landsat 8 images (DN already in range of (0, 255)) then
set scale_factor to 255.
}
\examples{
img_stk <- terra::rast(system.file("extdata", "BOA",
         "BOA_2022-12-11.tif", package = "rOPTRAM"))
vi <- calculate_vi(img_stk)
}
