% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilmoisture.R
\name{CalculateSoilMoisture}
\alias{CalculateSoilMoisture}
\title{Calculate soil moisture grid}
\usage{
CalculateSoilMoisture(img_date)
}
\arguments{
\item{string, }{img_date}
}
\value{
rast, soil moisture grid
}
\description{
Calculate soil moisture grid
}
\examples{
print("Running soilmoisture.R")
}
