% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_aoi}
\alias{check_aoi}
\title{Check aoi_file}
\usage{
check_aoi(aoi_file)
}
\arguments{
\item{aoi_file, }{string}
}
\value{
boolean, TRUE when file exists, and is spatial
}
\description{
Check that aoi file exists, and is a spatial file
}
\examples{
aoi_file <- system.file("extdata", "migda_aoi.gpkg")
check_aoi(aoi_file)
}
