% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_scihub_access}
\alias{check_scihub_access}
\title{Check that sen2r is installed}
\usage{
check_scihub_access(site = "http://scihub.copernicus.eu")
}
\arguments{
\item{site, }{string, Default is copernicus download site}
}
\value{
boolean, whether online and using sen2r is possible
}
\description{
Verify that sen2r is installed, and site is available.
Default download site: "https://scihub.copernicus.eu/"
}
\examples{
site <-  "http://scihub.copernicus.eu"
check_scihub_access(site)
}
