% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_scihub_access}
\alias{check_scihub_access}
\title{Check access to Sentinel Hub, using sen2r}
\usage{
check_scihub_access(scihub_user = NULL, scihub_pass = NULL)
}
\arguments{
\item{scihub_user, }{string, scihub username}

\item{scihub_pass, }{string, scihub password}
}
\value{
boolean, whether access to scihub using sen2r is possible
}
\description{
Verify that sen2r is installed,
and check credentials for access to Sentinel Hub
}
