% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_landsat.R
\name{crop_landsat_list}
\alias{crop_landsat_list}
\title{Crop List of Landsat Bands to AOI}
\usage{
crop_landsat_list(landsat_dir)
}
\arguments{
\item{landsat_dir, }{folder, contains the manually downloaded Landsat imagers}
}
\value{
cropped_list, list, paths to derived BOA stacks, cropped to study area
}
\description{
Utility function to prepare BOA bands cropped to Area of Interest
}
\examples{
\dontrun{
# landsat_dir is a directory containing the original Landsat image folders
landsat_dir <- "...enter full path here..."
cropped_landsat_list <- crop_landsat_list(landsat_dir)
}
}
