% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_landsat.R
\name{crop_landsat_list}
\alias{crop_landsat_list}
\title{Crop List of Landsat Bands to AOI}
\usage{
crop_landsat_list(landsat_dir)
}
\arguments{
\item{landsat_dir, }{string, directory containing the original downloaded Landsat imagery folders.}
}
\value{
cropped_list, list, paths to derived BOA stacks, cropped to study area
}
\description{
Utility function to prepare BOA bands cropped to Area of Interest.
}
\examples{
\dontrun{
cropped_landsat_list <- crop_landsat_list(landsat_dir)
}
}
