% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{exponential_coefficients}
\alias{exponential_coefficients}
\title{Utility Function to Prepare Exponential FittedEedges of Trapezoid}
\usage{
exponential_coefficients(df, output_dir)
}
\arguments{
\item{df, }{data.frame, values of VI and STR along edges of trapezoid}

\item{output_dir, }{string, path to save coefficients CSV file}
}
\value{
df, data.frame, the trapezoid line edge points
with fitted wet/dry values added
}
\description{
Called by \code{\link[rOPTRAM]{optram_wetdry_coefficients}}
to prepare exponential curve along trapezoid edges
Calculates the intercept and slope of both wet and dry edges
and updates the edges data.frame with these exp fitted values
Not exported
}
\note{
Three CSV files are saved:
\itemize{
\item the regressions coefficients,
\item the trapezoid edge points
\item RMSE of the fitted curve
}
}
\examples{
\dontrun{
  df <- read.csv(system.file("extdata", "trapezoid_edges.csv",
                package = "rOPTRAM"))
  output_dir <- tempdir()
  coeffs <- rOPTRAM::exponential_coefficients(df, output_dir)
  coeffs
}
}
