% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_str_dataframe.R
\name{NDVI_STR_DataFrame}
\alias{NDVI_STR_DataFrame}
\title{Prepare dataframe of pairs NDVI and STR values}
\usage{
NDVI_STR_DataFrame(STR_dir, NDVI_dir)
}
\arguments{
\item{STR_dir, }{path to STR raster file}

\item{NDVI_dir, }{path to NDVI raster file}
}
\value{
full_df, data.frame
}
\description{
Prepare dataframe of pairs NDVI and STR values
}
\examples{
print("Running ndvi_str_dataframe.R")

}
