% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram.R
\name{optram}
\alias{optram}
\title{Prepare Sentinel imagery for soil moisture prediction using OPTRAM.}
\usage{
optram(
  aoi_file,
  vidx = "NDVI",
  from_date,
  to_date,
  max_cloud = 15,
  scihub_user = NULL,
  scihub_pass = NULL,
  remove_safe = "yes",
  output_dir = tempdir()
)
}
\arguments{
\item{aoi_file, }{string, full path to polygon spatial file of area of interest}

\item{vidx, }{string, which VI to prepare, either 'NVDI' or 'SAVI'}

\item{from_date, }{string, the start of the date range, Formatted as "YYYY-MM-DD"}

\item{to_date, }{the end of the date range.}

\item{max_cloud, }{integer, maximum percent cloud cover, Default 15.}

\item{scihub_user, }{string, username on Copernicus Sentinel Hub}

\item{scihub_pass, }{string, password on Sentinel hub}

\item{remove_safe, }{string, "yes" or "no", whether to delete downloaded
SAFE directories after processing. Default "yes"}

\item{output_dir, }{string, directory to save coeffs_file, default is tempdir()}
}
\value{
coeffs_file, string, full path to saved CSV of wet-dry coefficients
}
\description{
The main wrapper function to download, and preprocess Sentinel 2 imagery
then prepare the OPTRAM trapezoid to derive slope and intercept
for wet and dry trapezoid lines.
See:
Sadeghi, M., Babaeian, E., Tuller, M., Jones, S.B., 2017.
The optical trapezoid model:
A novel approach to remote sensing of soil moisture
applied to Sentinel-2 and Landsat-8 observations.
Remote Sensing of Environment 198, 52–68,
https://doi.org/10.1016/j.rse.2017.05.041 .
}
\examples{
print("Running optram.R")
}
