% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram.R
\name{optram}
\alias{optram}
\title{Prepare Sentinel imagery for soil moisture prediction using OPTRAM.}
\usage{
optram(
  aoi_file,
  veg_index = "NDVI",
  from_date,
  to_date,
  max_cloud = 15,
  scihub_user = NULL,
  scihub_pass = NULL,
  remove_safe = "yes",
  S2_output_dir = tempdir(),
  data_output_dir = tempdir()
)
}
\arguments{
\item{aoi_file, }{string, full path to polygon spatial file of area of interest}

\item{veg_index, }{string, which index to use. Default "NDVI"}

\item{from_date, }{string, the start of the date range, Formatted as "YYYY-MM-DD"}

\item{to_date, }{the end of the date range.}

\item{max_cloud, }{integer, maximum percent cloud cover, Default 15.}

\item{scihub_user, }{string, username on Copernicus Sentinel Hub}

\item{scihub_pass, }{string, password on Sentinel hub}

\item{remove_safe, }{string, "yes" or "no", whether to delete downloaded
SAFE directories after processing. Default "yes"}

\item{S2_output_dir, }{string, directory to save downloaded S2
and the derived products, defaults to tempdir()}

\item{data_output_dir, }{string, path to save coeffs_file
and STR-VI data.frame, default is tempdir()}
}
\value{
coeffs_file, string, full path to saved CSV of wet-dry coefficients
}
\description{
The main wrapper function to download, and preprocess Sentinel 2 imagery
then prepare the OPTRAM trapezoid to derive slope and intercept
for wet and dry trapezoid lines.
See:
Sadeghi, M., Babaeian, E., Tuller, M., Jones, S.B., 2017.
The optical trapezoid model:
A novel approach to remote sensing of soil moisture
applied to Sentinel-2 and Landsat-8 observations.
Remote Sensing of Environment 198, 52–68,
https://doi.org/10.1016/j.rse.2017.05.041 .
}
\note{
Access to Copernicus Sentinel Hub requires registration.
If you have already registered, and saved your credentials
in the default "~/.sen2r/apihub.txt" file,
then you can leave the scihub_user and scihub_pass empty.
If your credentials are \strong{not} yet stored, then
\itemize{
\item register on the Scihub website:
\item https://scihub.copernicus.eu/userguide/SelfRegistration
\item enter your user and pass parameters in this function call,
and they will be stored into the default location.
}

Output can be separated:
Sentinel downloads and products are saved to S2_output_dir.
Data files (Trapezoid coefficients and STR-VI data) to data_output_dir
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2019-04-30"
aoi <- "inst/extdata/migda_9.gpkg"
coeffs <- optram(aoi,
                 from_date, to_date,
                 veg_index = c("SAVI"),
                 scihub_user = "userxxx", scihub_pass = "secretxyz"
                 )
}
}
