% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram.R
\name{optram}
\alias{optram}
\title{Prepare Sentinel Imagery for Soil Moisture Prediction Using OPTRAM.}
\usage{
optram(
  aoi,
  veg_index = "NDVI",
  from_date,
  to_date,
  max_cloud = 15,
  S2_output_dir = tempdir(),
  data_output_dir = tempdir(),
  remote = "scihub",
  vi_step = 0.005,
  trapezoid_method = c("linear", "exponential", "polynomial"),
  SWIR_band = c(11, 12)
)
}
\arguments{
\item{aoi, }{{sf} object, a POLYGON or MULTIPOLYGON, boundary of area of interest}

\item{veg_index, }{string, which index to use. Default "NDVI"}

\item{from_date, }{string, the start of the date range,
Formatted as "YYYY-MM-DD"}

\item{to_date, }{the end of the date range.}

\item{max_cloud, }{integer, maximum percent cloud cover, Default 15.}

\item{S2_output_dir, }{string, directory to save downloaded S2
and the derived products, defaults to tempdir()}

\item{data_output_dir, }{string, path to save coeffs_file
and STR-VI data.frame, default is tempdir()}

\item{remote, }{string, which Copernicus API to use,
one of "scihub", or "openeo". Default is "scihub"}

\item{vi_step, }{float, width of intervals along VI axis
default 0.005}

\item{trapezoid_method, }{string,
one of "linear", "exponential", "polynomial"
default "linear"
How to fit a curve to the values along trapezoid edges
See notes in optram_wetdry_coefficients()}

\item{SWIR_band, }{integer, either 11 or 12, determines which SWIR band to use}
}
\value{
rmse_df, data.frame, RMSE values of fitted trapezoid lines
the coefficients are also saved to a csv file in \code{data_output_dir}.
}
\description{
The main wrapper function to download, and preprocess Sentinel 2 imagery
then prepare the OPTRAM trapezoid to derive slope and intercept
for wet and dry trapezoid lines.
See:
Sadeghi, M., Babaeian, E., Tuller, M., Jones, S.B., 2017.
The optical trapezoid model:
A novel approach to remote sensing of soil moisture
applied to Sentinel-2 and Landsat-8 observations.
Remote Sensing of Environment 198, 52–68,
https://doi.org/10.1016/j.rse.2017.05.041 .
}
\note{
to download imagery. Please first install \code{gcloud} following instructions:
https://cloud.google.com/sdk/docs/install
for your operating system.
And be sure to initialize with you google username and password.

Output can be separated:
Sentinel downloads and products are saved to S2_output_dir.
Data files (Trapezoid coefficients and STR-VI data) to data_output_dir

Two SWIR wavelength bands are available in Sentinel-2:
1610 nanometer (nm) and 2190 nm.
The parameter \code{SWIR_bands } allows to choose which band is used in this model.
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2020-04-30"
aoi <- sf::st_read(system.file("extdata",
                              "lachish.gpkg", package = "rOPTRAM"))
rmse <- optram(aoi,
                 from_date, to_date,
                 veg_index = c("SAVI"),
                 trapezoid_method = "linear")
}
}
