% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram.R
\name{optram}
\alias{optram}
\title{Prepare Sentinel Imagery for Soil Moisture Prediction Using OPTRAM.}
\usage{
optram(
  aoi_file,
  veg_index = "NDVI",
  from_date,
  to_date,
  max_cloud = 15,
  remove_safe = "yes",
  timeperiod = "full",
  S2_output_dir = tempdir(),
  data_output_dir = tempdir(),
  remote = "gcloud"
)
}
\arguments{
\item{aoi_file, }{string, full path to polygon spatial file
of area of interest}

\item{veg_index, }{string, which index to use. Default "NDVI"}

\item{from_date, }{string, the start of the date range,
Formatted as "YYYY-MM-DD"}

\item{to_date, }{the end of the date range.}

\item{max_cloud, }{integer, maximum percent cloud cover, Default 15.}

\item{remove_safe, }{string, "yes" or "no", whether to delete downloaded
SAFE directories after processing. Default "yes"}

\item{timeperiod, }{string, either "full" for the whole date range,
or "seasonal" for only months specified, but over the full date range.}

\item{S2_output_dir, }{string, directory to save downloaded S2
and the derived products, defaults to tempdir()}

\item{data_output_dir, }{string, path to save coeffs_file
and STR-VI data.frame, default is tempdir()}

\item{remote, }{string, which Copernicus API to use,
one of "gcloud", "scihub", or "openeo"}
}
\value{
coeffs_file, string, full path to saved CSV of wet-dry coefficients
the coefficients are also saved to a csv file in \code{data_output_dir}.
}
\description{
The main wrapper function to download, and preprocess Sentinel 2 imagery
then prepare the OPTRAM trapezoid to derive slope and intercept
for wet and dry trapezoid lines.
See:
Sadeghi, M., Babaeian, E., Tuller, M., Jones, S.B., 2017.
The optical trapezoid model:
A novel approach to remote sensing of soil moisture
applied to Sentinel-2 and Landsat-8 observations.
Remote Sensing of Environment 198, 52–68,
https://doi.org/10.1016/j.rse.2017.05.041 .
}
\note{
to download imagery. Please first install \code{gcloud} following instructions:
https://cloud.google.com/sdk/docs/install
for your operating system.
And be sure to initialize with you google username and password.

Output can be separated:
Sentinel downloads and products are saved to S2_output_dir.
Data files (Trapezoid coefficients and STR-VI data) to data_output_dir

The \code{timeperiod} parameter allows to download either a "full" date range,
i.e. all images between the \code{from_date} and \code{to_date}, or alternatively,
"seasonal", where images are acquired for all years but only for the months
specified. So, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ if `from_date` = "2018-06-01` and `to_date` = '2020-09-20', then
 images for june through sept will be acquired for 2018, 2019, and 2020.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2020-04-30"
aoi_file <- system.file("extdata", "migda_aoi.gpkg", package = "rOPTRAM")
coeffs <- optram(aoi_file,
                 from_date, to_date,
                 veg_index = c("SAVI"),
                 timeperiod = "seasonal")
}
}
