% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_acquire_s2.R
\name{optram_acquire_s2}
\alias{optram_acquire_s2}
\title{Acquire Sentinel 2 images at a given location and date range}
\usage{
optram_acquire_s2(
  aoi_file,
  from_date,
  to_date,
  scihub_user = NULL,
  scihub_pass = NULL,
  max_cloud = 10,
  timeperiod = "full",
  output_dir = tempdir(),
  remove_safe = "yes",
  veg_index = "NDVI"
)
}
\arguments{
\item{aoi_file, }{string, full path to polygon spatial file of
boundary of area of interest}

\item{from_date, }{string, represents start of date range,
formatted as "YYYY-MM-DD"}

\item{to_date, }{string, end of date range, formatted as "YYYY-MM-DD"}

\item{scihub_user, }{string, username on Copernicus scihub}

\item{scihub_pass, }{string, password for access to Copernicus scihub}

\item{max_cloud, }{integer, maximum percent of cloud cover. Default 15.}

\item{timeperiod, }{string, either "full" for the whole date range,
or "seasonal" for only months specified, but over the full date range.}

\item{output_dir, }{string, path to save downloaded, and processed imagery}

\item{remove_safe, }{string, "yes" or "no":
whether to delete downloaded SAFE directories
after processing, default "yes"}

\item{veg_index, }{string, which index to prepare. Default "NDVI".
Can be "NDVI", "SAVI", "MSAVI", etc}
}
\value{
output_path, string, path to downloaded files
}
\description{
Use the \code{sen2r} package to acquire, preprocess and crop
Sentinel 2 satellite imagery.
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020)
sen2r: An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data
Computers & Geosciences, 139, 104473
doi: 10.1016/j.cageo.2020.104473
URL: https://sen2r.ranghetti.info/.
The package prepares vegetation indicies, cropped to an area of interest.
}
\note{
Access to Copernicus Sentinel Hub requires registration. If you have already
registered, and saved your credentials into the default file,
(typically: ~/.sen2r/apihub.txt)
then you can leave the scihub_user and scihub_pass empty.
If your credentials are \strong{not} yet stored, then
\itemize{
\item register on the Scihub website:
\item https://scihub.copernicus.eu/userguide/SelfRegistration
\item enter your user and pass parameters in this function call,
and they will be stored into the default location.
}

This function calls \code{sen2r()} from the {sen2r} package. This function
acquires Sentinel 2 imagery, clips to aoi,
and prepares multiband output rasters, save to the \code{output_dir}.
Only L2A (atmospherically corrected) images are acquired, and therefore:
\itemize{
\item only one NIR band, B08 is used. The lower res B8A is not used
\item furthermore, the cirrus band B09 is not relevant for BOA level
\item so band 10 is the SWIR refelctance at 1600 nm,
and band 11 is reflectance at 2200 nm.
}
}
\examples{
\dontrun{
from_date <- "2018-12-01"
to_date <- "2019-04-30"
aoi <- system.file("extdata", "migda_aoi.gpkg", package = 'rOPTRAM')
s2_file_list <- optram_acquire_s2(aoi,
                                 from_date, to_date,
                                 scihub_user = "userxxx",
                                 scihub_pass = "secretxyz"
                                 timeperiod = "full",
                                 veg_index = "MSAVI2")
}
}
