% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_soilmoisture.R
\name{optram_calculate_soil_moisture}
\alias{optram_calculate_soil_moisture}
\title{Calculate soil moisture grid}
\usage{
optram_calculate_soil_moisture(
  img_date,
  VI_dir,
  STR_dir,
  coeffs_file,
  output_dir = tempdir()
)
}
\arguments{
\item{img_date, }{string, image date of single Sentinel 2 acquisition
formatted as "YYYY-MM-DD"}

\item{VI_dir, }{string, full path to directory holding the VI rasters}

\item{STR_dir, }{string, full path to directory holding the STR rasters}

\item{coeffs_file, }{string, full path to csv file of wet-dry coefficients}

\item{output_dir, }{string, full path to output directory
for saving soil moisture raster}
}
\value{
rast, soil moisture grid
}
\description{
Prepare soil moisture grid from STR and VI images for a single date
using the derived slope and intercept coefficients
}
\examples{
print("Running optram_calculate_soil_moisture.R")
}
