% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_soilmoisture.R
\name{optram_calculate_soil_moisture}
\alias{optram_calculate_soil_moisture}
\title{Calculate soil moisture grid}
\usage{
optram_calculate_soil_moisture(img_date, coeffs_file, output_dir = tempfir())
}
\arguments{
\item{img_date:}{string, image date of Sentinel 2 acquisition}

\item{coeffs_file:}{string, full path to csv file of wet-dry coefficients}
}
\value{
rast, soil moisture grid
}
\description{
Calculate soil moisture grid
}
\examples{
print("Running optram_calculate_soil_moisture.R")
}
