% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_soilmoisture.R
\name{optram_calculate_soil_moisture}
\alias{optram_calculate_soil_moisture}
\title{Calculate soil moisture grid}
\usage{
optram_calculate_soil_moisture(
  img_date,
  VI_dir,
  STR_dir,
  coeffs_file,
  output_dir = tempdir()
)
}
\arguments{
\item{img_date, }{string, image date of single Sentinel 2 acquisition
formatted as "YYYY-MM-DD"}

\item{VI_dir, }{string, full path to directory holding the VI rasters}

\item{STR_dir, }{string, full path to directory holding the STR rasters}

\item{coeffs_file, }{string, full path to csv file of wet-dry coefficients}

\item{output_dir, }{string, full path to output directory
for saving soil moisture raster}
}
\value{
rast, soil moisture grid
}
\description{
Prepare soil moisture grid from STR and VI images for a single date
using the derived slope and intercept coefficients
}
\note{
This function is used after preparing the OPTRAM model coefficients with:
\code{optram_wetdryt_coefficients()}. Typically a new image date,
(that was not used for preparing the model),
will be referenced in the \code{img_date} parameter.
The resulting soil moisture raster is saved to \code{output_dir}.
}
\examples{
img_date <- "2023-03-11"
VI_dir <- system.file("extdata", "SAVI", package = "rOPTRAM")
STR_dir <- system_file("extdata", "STR", package = "rOPTRAM")
coeffs_file <- system_file("extdata", "coefficients.csv",
        package = "rOPTRAM")
SM <- optram_calculate_soil_moisture(img_date, VI_dir, STR_dir, coeffs_file)
}
