% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_landsat.R
\name{optram_landsat}
\alias{optram_landsat}
\title{Handle Satellite Imagery in Original Landsat Format}
\usage{
optram_landsat(
  landsat_dir,
  aoi_file,
  veg_index = "NDVI",
  LC_output_dir = tempdir(),
  data_output_dir = tempdir()
)
}
\arguments{
\item{landsat_dir, }{string, full path to containing folder of downloaded
(unzipped) Landsat data in original landsat format,
after atmospheric correction (L2A)}

\item{aoi_file, }{string, path to boundary polygon spatial file
of area of interest}

\item{veg_index, }{string, which VI to prepare, either 'NDVI' (default)
or 'SAVI' or 'MSAVI'}

\item{LC_output_dir, }{string, directory to save the derived products,
defaults to tempdir()}

\item{data_output_dir, }{string, path to save coeffs_file
and STR-VI data.frame, default is tempdir()}
}
\value{
rmse_df, data.frame, RMSE values of fitted trapezoid edges
}
\description{
Use this function to prepares vegetation index and
SWIR Transformed Reflectance (STR) rasters
when you have already downloaded landsat image files in advance.
This function assumes that atmospheric correction has been applied.
}
\note{
Unlike the \code{optram_acquire_s2} function, there is no implementation
for automatic download of Landsat images.
This function requires a directory, set in the \code{landsat_dir} parameter,
which contains the set of Landsat tiles downloaded manually
by the user, in advance.
This directory should contain folders of Landsat images, where each folder
consists of the individual Landsat bands as Geotiff files,
as well as the metadata files as downloaded from,
i.e. the USGS EarthExplorer (https://earthexplorer.usgs.gov/) website.
}
\examples{
\dontrun{
aoi_file <- system.file("extdata", "lachish.gpkg", package = "rOPTRAM")
optram_landsat(landsat_dir,  aoi_file,
               veg_index = 'SAVI',
               LC_output_dir = tempdir(), data_output_dir = tempdir())
}
}
