% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_ndvi_str.R
\name{optram_ndvi_str}
\alias{optram_ndvi_str}
\title{Prepare dataframe of pairs NDVI and STR values}
\usage{
optram_ndvi_str(STR_dir, NDVI_dir)
}
\arguments{
\item{STR_dir, }{path to STR raster files}

\item{NDVI_dir, }{path to NDVI raster files}
}
\value{
full_df, data.frame
}
\description{
Collect all pixel values of both vegetation index,
and Swir Transformed Reflectance, for a time series of images.
Prepare data.frame of all pairs of values
(as input for the \code{optram_wetdry_coefficients()} function)
}
\examples{
print("Running optram_ndvi_str.R")

}
