% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_ndvi_str.R
\name{optram_ndvi_str}
\alias{optram_ndvi_str}
\title{Prepare dataframe of pairs NDVI and STR values}
\usage{
optram_ndvi_str(
  STR_list,
  VI_list,
  output_dir = tempdir(),
  max_tbl_size = 5e+06,
  rm.low.vi = FALSE
)
}
\arguments{
\item{STR_list, }{list of paths to STR raster files}

\item{VI_list, }{list of paths to NDVI raster files}

\item{output_dir, }{string, path to save data.frames (in RDS format)}

\item{max_tbl_size, }{numeric, maximum size of NDVI-STR data.frame
default set to 5e+6}

\item{rm.low.vi, }{boolean, Should VI values <= 0 be removed
default FALSE. If set to TRUE then all VI <= 0 will be set to NA.}
}
\value{
full_df, data.frame with 5 columns: X,Y,Date,NDVI,STR
}
\description{
Collect all pixel values of both vegetation index,
and Swir Transformed Reflectance, for a time series of images.
Prepare data.frame of all pairs of values
(as input for the \code{optram_wetdry_coefficients()} function)
}
\note{
Use the \code{max_tbl_size} parameter to limit size of the NDVI-STR data.frame
With a large area of interest, and long time frame,
the number of data points can overrun the computation resources.
This parameter sets a total size of data.frame from the \code{max_tbl_size}
parameter, together with the number of image time slots in the time range.

In some cases (i.e. water surfaces) NDVI can have values
below zero. These pixels can be removed from the trapezoid
by setting \code{rm.low.vi} to TRUE.

The vegetation index column is named "NDVI" even though it can represent
other vegetgation indices, such as SAVI, or MSAVI.
}
\examples{
VI_list <- list.files(system.file("extdata", "SAVI"),
        pattern = ".tif$", full.names = TRUE)
STR_list <- list.files(system.file("extdata", "STR"),
        pattern = ".tif$", full.names = TRUE)
full_df <- optram_ndvi_str(STR_list, VI_list)
# Show structure of output data.frame
str(full_df)
}
