% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_ndvi_str.R
\name{optram_ndvi_str}
\alias{optram_ndvi_str}
\title{Prepare Dataframe with Pairs of NDVI and STR Values}
\usage{
optram_ndvi_str(STR_list, VI_list, output_dir = tempdir(), aoi = NULL)
}
\arguments{
\item{STR_list, }{list of paths to STR raster files}

\item{VI_list, }{list of paths to NDVI raster files}

\item{output_dir, }{string, path to save data.frames (in RDS format)}

\item{aoi, }{{sf} POLYGON or MULTIPOLYGON,
must have a numeric column named "ID" for coloring trapezoid points by features
Default NULL, (no coloring)}
}
\value{
full_df, data.frame with 7 columns: X,Y,Date,Month,NDVI,STR,Density
and optionally a 7th column with feature ID values from the AOI polygon.
The columns Month, Density, Feature_ID  can be used in
\code{plot_vi_str_cloud()} function to color the points in the scatter plot
in various ways.
}
\description{
Collect all pixel values of both vegetation index,
and Swir Transformed Reflectance, for a time series of images.
Prepare data.frame of all pairs of values
(as input for the \code{\link[rOPTRAM]{optram_wetdry_coefficients}} function)
}
\note{
Use the option \code{max_tbl_size} (see \code{\link[rOPTRAM]{optram_options}})
to limit size of the NDVI-STR data.frame.
With a large area of interest, and long time frame,
the number of data points can overrun the computation resources.
This parameter sets a total size of data.frame from the \code{max_tbl_size}
parameter, together with the number of image time slots in the time range.

In some cases (i.e. water surfaces) NDVI can have values
below zero. These pixels can be removed from the trapezoid
by setting \code{rm.low.vi}
option to TRUE.

The vegetation index column is named "NDVI" even though it can represent
other vegetation indices, such as SAVI, or MSAVI.
}
\examples{
VI_list <- list.files(system.file("extdata", "NDVI", package = "rOPTRAM"),
        pattern = ".tif$", full.names = TRUE)
STR_list <- list.files(system.file("extdata", "STR", package = "rOPTRAM"),
        pattern = ".tif$", full.names = TRUE)
full_df <- optram_ndvi_str(STR_list, VI_list)
# Show structure of output data.frame
str(full_df)
}
