% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{optram_options}
\alias{optram_options}
\title{Display or Set Package Options}
\usage{
optram_options(opt_name = NULL, opt_value = NULL, show_opts = TRUE)
}
\arguments{
\item{opt_name, }{string, one of the package options.
Default NULL}

\item{opt_value, }{string, numeric, or boolean, depending on opt_name.}

\item{show_opts, }{boolean, default TRUE, prints a list of all optram options
Default NULL}
}
\description{
{rOPTRAM} uses several package options.
This function displays the current defined options,
(showing the default options when the package is first loaded),
and allows users to set new values for each option.
}
\note{
When no new option name or value is specified,
a list of currently defined options is printed.

{rOPTRAM} defines the following options at startup
\tabular{lll}{
opt_name     \tab default  \tab other possible values \cr
---------------- \tab---------------\tab ------------------- \cr
veg_index    \tab "NDVI"  \tab "SAVI", "MSAVI" \cr
remote       \tab "scihub"   \tab "openeo" \cr
period       \tab "full"   \tab "seasonal" \cr
max_cloud    \tab 12 \tab between 0 and 100 \cr
vi_step      \tab 0.005  \tab usually between 0.01 and 0.001 \cr
trapezoid_method \tab "linear"  \tab "polynomial" or "exponential" \cr
SWIR_band    \tab 11    \tab 12 \cr
max_tbl_size \tab 1e+6  \tab depends on computer resources \cr
rm.low.vi    \tab FALSE \tab TRUE \cr
rm.hi.str    \tab FALSE \tab TRUE \cr
plot_colors  \tab "no"  \tab "no" = uniform green color for all points \cr
\tab  \tab "features" = points colored by aoi features \cr
\tab  \tab "density" = points colored by point density \cr
\tab  \tab "contours" = plots density contour lines \cr
\tab  \tab "months" = points colored by month of image date \cr
feature_col  \tab "ID" \tab string, name of \emph{numeric} column \cr
\tab  \tab that contains feature ID's for coloring plot \cr
edge_points  \tab TRUE  \tab FALSE, whether to add \cr
\tab  \tab the trapezoid edge points to the plot \cr
overwrite    \tab  FALSE \tab Set to TRUE to re-download already acquired S2 images \cr
}
}
\examples{
opts <- options()
optram_options()      # prints out list of current options
optram_options("SWIR_band", 12)
optram_options("veg_index", "SAVI")
optram_options("trapezoid_list", "exp")  # fails
options(opts)

}
