% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_other.R
\name{optram_prepare_other_vi_str}
\alias{optram_prepare_other_vi_str}
\title{Handle third party imagery with red, NIR and SWIR bands}
\usage{
optram_prepare_other_vi_str(
  img_dir,
  aoi,
  viname = "NDVI",
  output_dir = tempdir()
)
}
\arguments{
\item{img_dir, }{string, full path directory of downloaded images (assumes Geotif)}

\item{aoi, }{string, full path to polygon spatial file of area of interest}

\item{viname, }{string, which VI to prepare, 'NVDI', 'SAVI', etc.}

\item{output_dir, }{string, where to save Geotiff, default is tempdir()}
}
\value{
output_files, list, full paths to saved Geotif files
}
\description{
This function prepares vegetation index and STR rasters
from other (non-Sentinel) image files downloaded in advance.
}
\examples{
print("Running prepare_other_vi_str.R")
}
