% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_safe.R
\name{optram_safe}
\alias{optram_safe}
\title{Handle Sentinel Imagery in Original Copernicus SAFE Format}
\usage{
optram_safe(
  safe_dir,
  aoi,
  veg_index = "NDVI",
  S2_output_dir = tempdir(),
  overwrite = TRUE,
  data_output_dir = tempdir(),
  max_tbl_size = 5e+06,
  trapezoid_method = c("linear", "exponential", "polynomial"),
  vi_step = 0.01,
  SWIR_band = c(11, 12)
)
}
\arguments{
\item{safe_dir, }{string, full path to containing folder of downloaded
(unzipped) Sentinel 2 data in original SAFE format,
after atmospheric correction (L2A)}

\item{aoi, }{{sf} object, a POLYGON or MULTIPOLYGON of the AOI boundary}

\item{veg_index, }{string, which VI to prepare, either 'NDVI' (default)
or 'SAVI' or 'MSAVI'}

\item{S2_output_dir, }{string, directory to save the derived products,
defaults to tempdir()}

\item{overwrite, }{boolean, overwrite derived products
that already were created,
defaults to TRUE}

\item{data_output_dir, }{string, path to save coeffs_file
and STR-VI data.frame, default is tempdir()}

\item{max_tbl_size, }{numeric, maximum size of VI/STR dta.frame.
default is 5,000,000 rows}

\item{trapezoid_method, }{string,
one of "linear", "exponential", "polynomial"
how to prepare fitted model to trapezoid edge values
default "linear"}

\item{vi_step, }{numeric, width of slice along VI axis
to prepare edge values, default 0.01}

\item{SWIR_band, }{integer, either 11 or 12, determines which SWIR band to use}
}
\value{
rmse_df, data.frame, RMSE values of fitted trapezoid lines
}
\description{
Use this function to prepares vegetation index and
SWIR Transformed Reflectance (STR) rasters
when you have already downloaded Sentinel 2 image files in advance.
Unzip the downloaded Sentinel 2 files
and do not change the folder structure.
This function assumes that atmospheric correction has been applied.
i.e. by downloading Level L2A product or using the SNAP L2A_Process,
}
\note{
Use the  \code{max_tbl_size}  parameter to limit the total number of rows
in the VI-STR data.frame.
When the area of interest is large, or the time range of datasets is long,
the total size of the data.frame can grow
beyond the capacity of computation resources.
This parameter limits the size of the table
by sampling a number of data points from each time slot.
The sample size is determined based on  \code{max_tbl_size}
and the total number of time slots in the full time range.

Two SWIR bands are available in Sentinel-2: 1610 nanometer (nm) and 2190 nm.
The parameter \code{SWIR_band} allows to choose which band is used in this model.
}
\examples{
\dontrun{
aoi <- sf::st_read(system.file("extdata",
                  "lachish.gpkg", package = "rOPTRAM"))
safe_dir  <- "...enter directory of downloaded SAFE folders..."
rmse <- optram_safe(safe_dir,
                      aoi, veg_index = "NDVI", SWIR_band = 11)
# Plot trapezoid
full_df <- readRDS(system.file("extdata", "VI_STR_data.rds",
        package = "rOPTRAM"))
edges_df <- read.csv(system.file("extdata", "trapezoid_edges_lin.csv",
                        package = "rOPTRAM"))
pl <- plot_vi_str_cloud(full_df, edges_df, edge_points = TRUE)
pl
}
}
