% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_validate.R
\name{optram_validate}
\alias{optram_validate}
\title{Validate OPTRAM model with in situ data}
\usage{
optram_validate(
  data_file,
  model_output_dir,
  TDR_file,
  TDR_layer = NULL,
  output_dir = tempdir(),
  show_plot = FALSE
)
}
\arguments{
\item{data_file, }{string, full path to CSV file of in situ measurements. See Note below.}

\item{model_output_dir, }{string, path to directory of rasters from:
\verb{optram_calculate soil_moisture()} function}

\item{TDR_file, }{string, full path to spatial file of TDR sensor.
Either geopackage, shapefile, or other spatial data format.}

\item{TDR_layer, }{string, layer name of TDR location
(required only if the spatial file is geopackage).}

\item{output_dir, }{string, where to save data.frame of soil moisture values
and (optional) plot.}

\item{show_plot, }{boolean, whether to prepare regression plot, default FALSE}
}
\value{
sm_df, data.frame of soil moisture values, in situ and predicted
This data is also saved to and *.rds file.
}
\description{
Validation of the OPTRAM model is performed by reading a CSV file
of ground truth soil moisture data, typically from TDR sensors.
The CSV data file represents measurements over a time series, for one location.
The predicted soil moisture values from the model are extracted at the sensor location
from a time series of OPTRAM model prediction rasters.
(derived from the \verb{optram_calculate soil_miosture()} function)
The ground truth measurements are correlated with the model predictions
and an optional scatterplot with regression line are plotted.
}
\note{
The data file should contain two columns:
the first column must be date time in ISO format: "YYYY-MM-DD HH:MM:SSZ"
The second column must be the Volumetric Water Content in percent
(all other columns are ignored)
}
\examples{
#' print("Running optram_validate.R")
}
