% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{optram_wetdry_coefficients}
\alias{optram_wetdry_coefficients}
\title{Derive coefficients of slope and intercept}
\usage{
optram_wetdry_coefficients(full_df, step = 0.001)
}
\arguments{
\item{full_df, }{data.frame of STR and NDVI values}

\item{step, }{float}
}
\value{
list of float, coefficients of wet-dry trapezoid
}
\description{
Derive slope and intercept coefficients
for both wet and dry trapezoid lines.
Write coefficients to a CSV file (as input to \code{optram_soilmoisture()} function)
}
\examples{
print("Running optram_wetdry_coefficients.R")

}
