% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{optram_wetdry_coefficients}
\alias{optram_wetdry_coefficients}
\title{Derive coefficients of slope and intercept}
\usage{
optram_wetdry_coefficients(
  full_df,
  aoi_file,
  output_dir = tempdir(),
  step = 0.01,
  save_plot = TRUE
)
}
\arguments{
\item{full_df, }{data.frame of STR and NDVI values}

\item{aoi_file, }{string, added to title of plot
(Can be path to AOI file, then the file name is used in plot title)}

\item{output_dir, }{string, directory to save coefficients CSV file}

\item{step, }{float}

\item{save_plot, }{boolean, If TRUE (default) save scatterplot to output_dir}
}
\value{
coeffs, list of float, coefficients of wet-dry trapezoid
}
\description{
Derive slope and intercept coefficients
for both wet and dry trapezoid lines.
Write coefficients to a CSV file
(as input to \code{optram_soilmoisture()} function)
}
\note{
The vegetation index column is named "NDVI" even though it can represent
other vegetgation indices, such as SAVI, or MSAVI.
}
\examples{
aoi_file <- "Test"
full_df <- readRDS(system.file("extdata",
              "VI_STR_data.rds",
              package = "rOPTRAM"))
coeffs <- optram_wetdry_coefficients(full_df, aoi_file)
print(coeffs)
}
