% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{plot_cloud_exponential}
\alias{plot_cloud_exponential}
\title{Scatterplot, Exponential Fitted Curve}
\usage{
plot_cloud_exponential(pl_base, output_dir, aoi_name)
}
\arguments{
\item{pl_base, }{ggplot object, additional lines are added to base plot}

\item{aoi_name, }{string, added to plot title}

\item{plot_df, }{data.frame, original plot data, needed for Exponential fitting}

\item{coeffs, }{data.frame, 4 trapezoid coefficients}
}
\value{
ggplot object
}
\description{
Plot of VI/STR points, with exponential fitted curve
}
\note{
This function is called from \code{plot_vi_str_cloud()}
}
\examples{
\dontrun{
plot_df <- readRDS(system.file("extdata",
                               "VI_STR_data.RDS",package = "rOPTRAM")
coeffs_file <- system.file("extdata", "coefficients.csv",
        package = "rOPTRAM")
coeffs <- read.csv(coeffs_file)
pl <- plot_cloud_exponential(plot_df, coeffs)
}
}
