% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{plot_cloud_linear}
\alias{plot_cloud_linear}
\title{Scatterplot, Linear Fitted Curve}
\usage{
plot_cloud_linear(pl_base, aoi_name)
}
\arguments{
\item{pl_base, }{ggplot object, additional lines are added to base plot}

\item{aoi_name, }{string, added to plot title}

\item{output_dir, }{string, path to directory where trapezoid edges were saved}
}
\value{
ggplot object
}
\description{
Plot of VI/STR points, with linear fitted curve
}
\note{
This function is called from \code{plot_vi_str_cloud()}
}
\examples{
\dontrun{
plot_df <- readRDS(system.file("extdata",
                               "VI_STR_data.RDS",package = "rOPTRAM")
coeffs_file <- system.file("extdata", "coefficients_lin.csv",
        package = "rOPTRAM")
coeffs <- read.csv(coeffs_file)
pl <- plot_cloud_linear(plot_df, coeffs)
}
}
