% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ndvi_str_cloud.R
\name{Plot_STR_NDVI_Cloud}
\alias{Plot_STR_NDVI_Cloud}
\title{Create scatter plot of STR-NDVI point cloud,
with dry and wet trapezoid lines}
\usage{
Plot_STR_NDVI_Cloud(full_df, coeffs)
}
\arguments{
\item{full_df, }{data.frame of NDVI and STR pixel values}

\item{coeffs, }{list of floats, the slope and intercept
of wet and dry regression lines}
}
\value{
None
}
\description{
Create scatter plot of STR-NDVI point cloud,
with dry and wet trapezoid lines
}
\examples{
print("Running plot_ndvi_str_cloud.R")

}
