% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{plot_vi_str_cloud}
\alias{plot_vi_str_cloud}
\title{Create Scatter Plot of STR-NDVI Data Points,}
\usage{
plot_vi_str_cloud(full_df, edges_df, edge_points = FALSE)
}
\arguments{
\item{full_df, }{data.frame of NDVI and STR pixel values}

\item{edges_df, }{data.frame, points along the wet/dry edges for trapezoid}

\item{edge_points, }{boolean, whether to add to the plot the
linear regression points that were used to derive coefficients.
default FALSE}
}
\value{
ggplot object
}
\description{
Plot STR-NDVI scatterplot to show dry and wet trapezoid lines
over scatterplot of multi-temporal STR and NDVI pixel values
}
\note{
Points in scatter plot can be colored in various ways, depending on the
\code{plot_colors} option, as set in \code{optram_options()}
If "plot_colors = none" all points are colored green.
If "plot_colors = density" points are colored by point density on the plot.
If "plot_colors = contour" points are colored green and density contour lines
are overlayed
If "plot_colors = feature" points are colored by some feature ID
in the original AOI polygon.
If "plot_colors = month" points are colored by the month of image acquisition.
}
\examples{
aoi_name <- "Soil Moisture AOI"
optram_options("trapezoid_method", "polynomial")
full_df <- readRDS(system.file("extdata", "VI_STR_data.rds",
        package = "rOPTRAM"))
edges_df <- read.csv(system.file("extdata", "trapezoid_edges_lin.csv",
                        package = "rOPTRAM"))
pl <- plot_vi_str_cloud(full_df, edges_df)
pl + ggplot2::ggtitle(paste("Trapezoid plot for:", aoi_name))
pl

}
