% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{plot_vi_str_cloud}
\alias{plot_vi_str_cloud}
\title{Create Scatter Plot of STR-NDVI Data Points,}
\usage{
plot_vi_str_cloud(
  full_df,
  coeffs,
  aoi_name,
  output_dir = tempdir(),
  trapezoid_method = c("linear", "exponential", "polynomial"),
  edges_points = FALSE
)
}
\arguments{
\item{full_df, }{data.frame of NDVI and STR pixel values}

\item{coeffs, }{list of floats, the slope and intercept
of wet and dry regression lines}

\item{aoi_name, }{string, used in plot title}

\item{output_dir, }{string, directory to save plot png file.}

\item{trapezoid_method, }{string, how to plot trapezoid line.
either "linear" or "exponential", default is "linear"}

\item{edges_points, }{boolean, whether to add to the plot the
linear regression points that were used to derive coefficients.
default FALSE}
}
\value{
None
}
\description{
Plot STR-NDVI scatterplot to show dry and wet trapezoid lines
over scatterplot of multi-temporal STR and NDVI pixel values
}
\examples{
aoi_name <- "Test"
full_df <- readRDS(system.file("extdata", "VI_STR_data.rds",
        package = "rOPTRAM"))
coeffs <- read.csv(system.file("extdata", "coefficients_lin.csv",
        package = "rOPTRAM"))
plot_vi_str_cloud(full_df, coeffs, aoi_name)
coeffs <- read.csv(system.file("extdata", "coefficients_pol.csv",
        package = "rOPTRAM"))
plot_vi_str_cloud(full_df, coeffs, aoi_name,
                    trapezoid_method = "polynomial")

}
