% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optram_wetdry_coefficients.R
\name{plot_vi_str_cloud}
\alias{plot_vi_str_cloud}
\title{Create Scatter Plot of STR-NDVI Data Points,}
\usage{
plot_vi_str_cloud(
  full_df,
  edges_df,
  trapezoid_method = c("linear", "exponential", "polynomial"),
  edge_points = FALSE
)
}
\arguments{
\item{full_df, }{data.frame of NDVI and STR pixel values}

\item{edges_df, }{data.frame, points along the wet/dry edges for trapezoid}

\item{trapezoid_method, }{string, how to plot trapezoid line.
either "linear" or "exponential", default is "linear"}

\item{edge_points, }{boolean, whether to add to the plot the
linear regression points that were used to derive coefficients.
default FALSE}
}
\value{
ggplot object
}
\description{
Plot STR-NDVI scatterplot to show dry and wet trapezoid lines
over scatterplot of multi-temporal STR and NDVI pixel values
}
\examples{
aoi_name <- "Soil Moisture AOI"
full_df <- readRDS(system.file("extdata", "VI_STR_data.rds",
        package = "rOPTRAM"))
edges_df <- read.csv(system.file("extdata", "trapezoid_edges_lin.csv",
                        package = "rOPTRAM"))
pl <- plot_vi_str_cloud(full_df, edges_df)
pl + ggplot2::ggtitle(paste("Trapezoid plot for:", aoi_name))
pl

edges_df <- read.csv(system.file("extdata", "trapezoid_edges_poly.csv",
                        package = "rOPTRAM"))
pl <- plot_vi_str_cloud(full_df, edges_df,
                    trapezoid_method = "polynomial")
pl + ggplot2::ggtitle(paste("Trapezoid plot for:", aoi_name))
pl

}
