% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid_utilities.R
\name{polynomial_coefficients}
\alias{polynomial_coefficients}
\title{Utility Function to prepare polynomial fitted edges of trapezoid}
\usage{
polynomial_coefficients(df, output_dir)
}
\arguments{
\item{df, }{data.frame, values of VI and STR along edges of trapezoid}

\item{output_dir, }{string, path to save coefficients CSV file}
}
\value{
coeffs, data.frame of 3 coefficients
for both wet and dry edges
}
\description{
Called by \code{optram_wetdry_coefficients()}
to prepare second order polynomial curve along trapezoid edges
Calculates six coefficients:  intercept (alpha)  of both wet and dry edges
and first and second order coefficients (beta), as in
$STR = alpha + beta_1 * VI + beta_2 * VI^2$
and updates the edges data.frame with these polynomila values fitted values
}
\examples{
\dontrun{
  df <- read.csv(system.file("extdata", "trapezoid_edges.csv",
                  package = "rOPTRAM"))
  output_dir <- tempdir()
  coeffs <- rOPTRAM::polynomial_coefficients(df, output_dir)
  coeffs
 }
}
