% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOPTRAM-package.R
\docType{package}
\name{rOPTRAM-package}
\alias{rOPTRAM}
\alias{rOPTRAM-package}
\title{rOPTRAM: Derive Soil Moisture Using the OPTRAM Algorithm}
\description{
The OPtical TRapezoid Model (OPTRAM) derives soil moisture based on the linear relation between a vegetation index and Land Surface Temperature (LST). The Short Wave Infra-red (SWIR) band is used as a proxy for LST. See: Sadeghi, M. et al., 2017. \url{https://doi.org/10.1016/j.rse.2017.05.041} .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/rOPTRAM.git}
  \item Report bugs at \url{https://github.com/ropensci/rOPTRAM/issues}
}

}
\author{
\strong{Maintainer}: Micha Silver \email{silverm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-1128-1325}{ORCID})

Other contributors:
\itemize{
  \item Arnon Karnieli \email{karnieli@bgu.ac.il} (\href{https://orcid.org/0000-0001-8065-9793}{ORCID}) [contributor, funder]
  \item Ron Beiden \email{beiden@post.bgu.ac.il} [contributor]
  \item Natalya Panov \email{npanov@bgu.ac.il} [contributor]
  \item Zhe Dong \email{dongzhe053@gmail.com} (\href{https://orcid.org/0009-0007-5298-8355}{ORCID}) [contributor]
  \item Lauren O'Brien \email{obrlsoilau@gmail.com} [reviewer]
  \item Harry Eslick \email{harry@wineandtruffle.com.au} [reviewer]
}

}
\keyword{internal}
