% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetdry_coefficients.R
\name{WetDry_Coefficients}
\alias{WetDry_Coefficients}
\title{Derive coefficients of slope and intercept
for both wet and dry trapezoid lines}
\usage{
WetDry_Coefficients(full_df, step = 0.001)
}
\arguments{
\item{full_df, }{data.frame of STR and NDVI values}

\item{step, }{float}
}
\value{
list of float, coefficients of wet-dry trapezoid
}
\description{
Derive coefficients of slope and intercept
for both wet and dry trapezoid lines
}
\examples{
print("Running wetdry_coefficients.R")

}
